maint: fix printf double conversion for 972 and 973

+++ src/tws_client.cpp
@@ -178,8 +178,8 @@ void TWSClient::cancelMktData ( int tickerId )
 void TWSClient::placeOrder ( int id, const IB::Contract &contract,
 	const IB::Order &order )
 {
-	DEBUG_PRINTF("PLACE_ORDER %d '%s' %ld '%s' %g '%s' %ld",
-		id, order.orderType.c_str(), order.totalQuantity, order.action.c_str(),
+	DEBUG_PRINTF("PLACE_ORDER %d '%s' %g '%s' %g '%s' %ld",
+		id, order.orderType.c_str(), (double)order.totalQuantity, order.action.c_str(),
 		order.lmtPrice, contract.symbol.c_str(), contract.conId );
 
 	ePosixClient->placeOrder( id, contract, order );
